requirements.txt
tensorflow>=2.13.0
streamlit>=1.24.0
numpy>=1.24.3
pandas>=2.0.3
pillow>=9.5.0
plotly>=5.15.0
opencv-python>=4.8.0
matplotlib>=3.7.1
seaborn>=0.12.2
scikit-learn>=1.3.0